third_party/blink/web_tests/fast/multicol/newmulticol/balance6-expected.txt

PASS mc.offsetHeight is 33
PASS mc.scrollWidth is 500
PASS successfullyParsed is true

TEST COMPLETE


There should be no scrollbar on this page.
